Fixed bug where product image model in client fell out of sync with db
Occured when a competing model took over primary property for that
product. symptom was unclicking 'active' on a view of image after
it was previously primary, made the image become primary again on the view
(its primary field was still set to true even though another image
had reclaimed it on the server-side).